sa/requirements.txt
# Install necessary packages (GPU version with CUDA 12.4)
regex
pandas
numpy
tqdm
openpyxl
soynlp
jieba
mecab-ko-msvc
mecab-ko-dic-msvc
# torch==2.6.0 # 이전 라인 주석 처리 또는 삭제
# torchvision==0.21.0 # 이전 라인 주석 처리 또는 삭제
# PyTorch (GPU - CUDA 12.4)
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu124
torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu124
# torchaudio도 필요하다면 추가 (예: torchaudio --index-url https://download.pytorch.org/whl/cu124)
transformers
FlagEmbedding
openai
sentence-transformers